pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-agent"
version = "0.1.0"
description = "A game writing assistant workbench (Streamlit + tool/skill agent)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

# 运行时依赖：与根目录 requirements.txt 保持一致
dependencies = [
  "streamlit==1.55.0",
  "httpx==0.28.1",
  "python-dotenv==1.2.2",
  "certifi==2026.2.25",
]

[project.optional-dependencies]
dev = [
  "modelscope-agent==0.7.2",
  "openai==2.29.0",
  "langchain-core==1.2.21",
  "truststore==0.10.4",
  "pytest==8.4.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["venv*", "venv_skills*", "data*", "logs*", "tmp*", "skills*"]

[tool.pytest.ini_options]
markers = [
  "integration: tests that hit network/external services",
]

